feat: redesign `read_msfragger()`
Aggregating PSMs is an essential step for any downstream analysis, and it's the prerequisite to merging different MSFragger result files. Therefore, after rethinking, we decided to incorporate this step to every functions in `glyread`. We've already implement aggregation for pGlycoQuant results in the commits before. In this commit, we redesign `read_msfragger()` to support reading all PSM files from a MSFragger-Glyco result folder, aggregating to the glycopeptide level, and merge the results. It now directly accepts a folder path as its first argument, and return a typical `glyexp::experiment()` as others. No further manual aggregation and merging is needed anymore.