fix: optimize the logic of protein inference
Protein inference for pGlyco3 result was performed on the PSM level. As one peptide may appears multiple times due to different charges and modifications (including glycans), the result might be wrong. The bug has been fixed that the protein inference is performed after taking the unique combination of peptides and proteins.